+++ backend/src/main/java/com/hexavolt/backend/dto/PowerDTO.java
@@ -3,15 +3,15 @@ package com.hexavolt.backend.dto;
 import java.math.BigDecimal;
 
 public class PowerDTO {
-    private Integer id;
+    private Long id;
     private BigDecimal kvaPower;
 
-    public PowerDTO(Integer id, BigDecimal kvaPower) {
+    public PowerDTO(Long id, BigDecimal kvaPower) {
         this.id = id;
         this.kvaPower = kvaPower;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
