+++ backend/src/main/java/com/hexavolt/backend/dto/UserMeDTO.java
@@ -1,4 +1,4 @@
 package com.hexavolt.backend.dto;
 
-public record UserMeDTO(Integer id, String email) {}
-
+public record UserMeDTO(Long id, String email) {
+}
