+++ backend/src/main/java/com/hexavolt/backend/dto/UserResponseDTO.java
@@ -3,14 +3,14 @@ package com.hexavolt.backend.dto;
 import java.time.LocalDate;
 
 public record UserResponseDTO(
-  Integer id,
-  String firstName,
-  String lastName,
-  String address,
-  String postalCode,
-  String phone,
-  LocalDate birthdate,
-  Integer cityId,
-  String email,
-  Boolean emailIsValid
-) {}
+    Long id,
+    String firstName,
+    String lastName,
+    String address,
+    String postalCode,
+    String phone,
+    LocalDate birthdate,
+    Long cityId,
+    String email,
+    Boolean emailIsValid) {
+}
