+++ backend/src/main/java/com/hexavolt/backend/entity/City.java
@@ -10,13 +10,13 @@ import jakarta.persistence.UniqueConstraint;
 
 @Entity
 @Table(name = "city", uniqueConstraints = {
-    @UniqueConstraint(columnNames = "name")
+        @UniqueConstraint(columnNames = "name")
 })
 public class City {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 100, nullable = false)
     private String name;
@@ -30,11 +30,11 @@ public class City {
     }
 
     // --- Getters and Setters ---
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
