+++ backend/src/main/java/com/hexavolt/backend/entity/WeeklySchedule.java
@@ -17,7 +17,7 @@ public class WeeklySchedule {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "day_of_week", referencedColumnName = "id")
@@ -34,9 +34,11 @@ public class WeeklySchedule {
     private LocalTime endTime;
 
     // --- Constructors ---
-    public WeeklySchedule() {}
+    public WeeklySchedule() {
+    }
 
-    public WeeklySchedule(DayOfWeek dayOfWeek, ChargingStation chargingStation, LocalTime startTime, LocalTime endTime) {
+    public WeeklySchedule(DayOfWeek dayOfWeek, ChargingStation chargingStation, LocalTime startTime,
+            LocalTime endTime) {
         this.dayOfWeek = dayOfWeek;
         this.chargingStation = chargingStation;
         this.startTime = startTime;
@@ -45,11 +47,11 @@ public class WeeklySchedule {
 
     // --- Getters and Setters ---
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
