+++ backend/src/main/java/com/hexavolt/backend/repository/ChargingStationRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hexavolt.backend.entity.ChargingStation;
 
-public interface ChargingStationRepository extends JpaRepository<ChargingStation, Integer> {
-  List<ChargingStation> findByLocationId(Integer locationId);
+public interface ChargingStationRepository extends JpaRepository<ChargingStation, Long> {
+  List<ChargingStation> findByLocationId(Long locationId);
   // Page<ChargingStation> findByOwner_Id(Integer ownerId, Pageable pageable);
   // Page<ChargingStation> findByActiveTrue(Pageable pageable);
